requirements.txt
pandas
python-docx
streamlit
requests